requirements.txt
earthengine-api>=0.1.370
opencage>=2.2.0
requests>=2.31.0
numpy>=1.24.0
rasterio>=1.3.8
scikit-learn>=1.3.0
tensorflow>=2.13.0
Pillow>=10.0.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
geopy>=2.3.0
supabase>=2.0.0
psycopg2-binary>=2.9.9

